test: replace waffletest with vitest across six environments
