+++ .github/workflows/test.yml
@@ -27,9 +27,9 @@ jobs:
             ${{ runner.os }}-
       - name: Install dependencies
         if: steps.cache-node-modules.outputs.cache-hit != 'true'
-        run: npx playwright install && npm ci
+        run: npm ci
       - name: Install Playwright Browsers
-        run: npx playwright install --with-deps
+        run: npx playwright install --with-deps chromium firefox webkit
       - name: Run browser tests
         run: npm run test:browser
 
@@ -62,7 +62,7 @@ jobs:
         if: steps.cache-node-modules.outputs.cache-hit != 'true'
         run: npm ci
       - name: Run tests
-        run: npm run test:node
+        run: npm test
 
   testTypes:
     name: 'Test: Types'
@@ -105,3 +105,38 @@ jobs:
         run: bun install --frozen-lockfile
       - name: Run tests
         run: npm run test:bun
+
+  # Known-failing environments. These run so the failures stay visible, but they do not gate the
+  # rest of the matrix: happy-dom reports the test server's requests as cross-origin and blocks
+  # them, and workerd has its own set of issues (a null `MessageEvent.origin`, on-handlers firing
+  # twice, and handler ordering that differs from every other runtime). Both are being worked
+  # through separately - drop `continue-on-error` once they are green.
+  testHappyDom:
+    name: 'Test: happy-dom (allowed to fail)'
+    timeout-minutes: 15
+    runs-on: ubuntu-latest
+    continue-on-error: true
+    steps:
+      - uses: actions/checkout@v7
+      - uses: actions/setup-node@v7
+        with:
+          node-version: 24
+      - name: Install dependencies
+        run: npm ci
+      - name: Run tests
+        run: npm run test:happy-dom
+
+  testWorkerd:
+    name: 'Test: Cloudflare Workers (allowed to fail)'
+    timeout-minutes: 15
+    runs-on: ubuntu-latest
+    continue-on-error: true
+    steps:
+      - uses: actions/checkout@v7
+      - uses: actions/setup-node@v7
+        with:
+          node-version: 24
+      - name: Install dependencies
+        run: npm ci
+      - name: Run tests
+        run: npm run test:workerd
