--- test/fixtures.ts
@@ -1,4 +1,0 @@
-export const unicodeLines = [
-  '🦄 are cool. 🐾 in the snow. Allyson Felix, 🏃🏽‍♀️ 🥇 2012 London!',
-  'Espen ♥ Kokos',
-]
