+++ test/helpers/fixtures.ts
@@ -1,0 +1,5 @@
+/** Typed as a tuple so the indexed reads in the server and tests are known to be present. */
+export const unicodeLines: [string, string] = [
+  '🦄 are cool. 🐾 in the snow. Allyson Felix, 🏃🏽‍♀️ 🥇 2012 London!',
+  'Espen ♥ Kokos',
+]
