--- test/waffletest/index.ts
@@ -1,2 +1,0 @@
-export * from './runner.ts'
-export * from './types.ts'
